requirements.txt
tensorflow
streamlit
pillow
numpy
tensorflow-hub
